fix: lens-debate fixes — rename branch→branchMode, move branchFromAwareness to data layer, remove dead code
Lens-debate consensus (hickey + lowy) applied 7 fixes:

1. Rename GitStatusUpdate.branch → branchMode (one name per concept)
2. Align formatGitStatusJson args with snapshotGitStatus return shape
3. Delete dead BranchInfo interface (exported, never consumed)
4. Remove dead log parameter from startGitStatus/requery signatures
5. Inline statusGlyph (no-op identity) — status code IS the glyph
6. Extract shared validateEndpointFlags helper (deduplicate bin.ts)
7. Move branchFromAwareness from render to data layer (restore
   data-never-imports-from-projection symmetry)
8. Remove .catch(() => null) from snapshotGitStatus git.getStatus calls
   (fail-fast: query errors surface, not collapse to null)
